move legacy config v1 to cmd
